requirements.txt
psycopg2-binary==2.8.6
configparser==4.0.2
Flask==1.1.1
simplejson==3.17.2
